+++ app/currencies/application/actions/ImportRates.php
@@ -3,39 +3,37 @@
 namespace app\currencies\application\actions;
 
 use app\currencies\application\forms\CurrencyForm;
-use app\currencies\application\providers\ProviderInterface;
+use app\currencies\application\services\RateServiceInterface;
 use app\currencies\domain\entities\Currency;
-use app\shared\application\exceptions\InvalidCallException;
+use app\shared\application\exceptions\UnexpectedValueException;
 
 class ImportRates extends BaseAction implements ImportRatesInterface
 {
     /**
-     * @param ProviderInterface $currencyRateProvider
+     * @param RateServiceInterface $rateService
      * @param CreateOrUpdateCurrencyInterface $createOrUpdateCurrency
+     * @param string $sourceCurrency
+     * @param string $targetCurrency
      */
     public function __construct(
-        private readonly ProviderInterface $currencyRateProvider,
+        private readonly RateServiceInterface $rateService,
         private readonly CreateOrUpdateCurrencyInterface $createOrUpdateCurrency,
+        private readonly string $sourceCurrency,
+        private readonly string $targetCurrency
     ) {
     }
 
     /**
      * @return Currency[]
-     * @throws InvalidCallException
+     * @throws UnexpectedValueException
      */
     public function execute(): array
     {
-        $rates = $this->currencyRateProvider->getActualRates();
-        if (empty($rates)) {
-            throw new InvalidCallException('Currency rate provider return empty');
-        }
-
-        $currencies = [];
-        foreach ($rates as $dto) {
-            $currencies[] = $this->createOrUpdateCurrency->execute(
-                new CurrencyForm($dto->getCurrency(), $dto->getRate())
-            );
-        }
-        return $currencies;
+        $rateDto = $this->rateService->getRate($this->sourceCurrency, $this->targetCurrency);
+        return [
+            $this->createOrUpdateCurrency->execute(
+                new CurrencyForm($rateDto->getCurrency(), $rateDto->getRate())
+            )
+        ];
     }
 }