+++ app/currencies/application/services/RateServiceInterface.php
@@ -1,15 +1,17 @@
 <?php
 
-namespace app\currencies\application\providers;
+namespace app\currencies\application\services;
 
 use app\currencies\application\dto\CurrencyProviderDto;
+use app\shared\application\exceptions\UnexpectedValueException;
 
-interface ProviderInterface
+interface RateServiceInterface
 {
     /**
      * @param string $sourceCurrency
      * @param string $targetCurrency
      * @return CurrencyProviderDto
+     * @throws UnexpectedValueException
      */
     public function getRate(string $sourceCurrency, string $targetCurrency): CurrencyProviderDto;
 }