+++ app/currencies/infrastructure/providers/CoinbaseProvider.php
@@ -6,7 +6,6 @@ use app\currencies\application\dto\CurrencyProviderDto;
 use app\currencies\application\providers\ProviderInterface;
 use app\shared\application\exceptions\RemoteServiceException;
 use app\shared\application\exceptions\UnexpectedValueException;
-use Exception;
 use GuzzleHttp\Client as HttpClient;
 use Psr\Http\Message\ResponseInterface;
 use Throwable;
@@ -17,28 +16,26 @@ class CoinbaseProvider extends BaseProvider implements ProviderInterface
     /**
      * CoinbaseProvider constructor.
      * @param HttpClient $client
-     * @param string $baseCurrency
-     * @param string $importCurrency
      */
     public function __construct(
         private readonly HttpClient $client,
-        private readonly string $baseCurrency = "USD",
-        private readonly string $importCurrency = "USD",
     ) {
     }
 
 
     /**
-     * @return CurrencyProviderDto[]
+     * @param string $sourceCurrency
+     * @param string $targetCurrency
+     * @return CurrencyProviderDto
      * @throws RemoteServiceException
      * @see https://docs.cdp.coinbase.com/sign-in-with-coinbase/docs/api-prices/#get-buy-price
      */
-    public function getActualRates(): array
+    public function getRate(string $sourceCurrency, string $targetCurrency): CurrencyProviderDto
     {
-        $url = sprintf("/v2/prices/%s-%s/buy", $this->baseCurrency, $this->importCurrency);
+        $url = sprintf("/v2/prices/%s-%s/buy", $sourceCurrency, $targetCurrency);
         try {
             $response = $this->client->get($url);
-            return $this->processResponse($response);
+            return $this->processResponse($response, $targetCurrency);
         } catch (RemoteServiceException $e) {
             throw $e;
         } catch (Throwable $e) {
@@ -49,19 +46,21 @@ class CoinbaseProvider extends BaseProvider implements ProviderInterface
 
     /**
      * @param ResponseInterface $response
-     * @return CurrencyProviderDto[]
-     * @throws Exception
+     * @param string $targetCurrency
+     * @return CurrencyProviderDto
+     * @throws RemoteServiceException
+     * @throws UnexpectedValueException
      */
-    protected function processResponse(ResponseInterface $response): array
+    protected function processResponse(ResponseInterface $response, string $targetCurrency): CurrencyProviderDto
     {
         $this->checkStatusCode($response, 200);
+
+        /** @var array{data?: array<array{amount: float}>} $body */
         $body = $this->parseJsonBody($response);
         if (empty($body['data']['amount'])) {
             throw new UnexpectedValueException('Bad conversion rate');
         }
 
-        return [
-            new CurrencyProviderDto($this->importCurrency, round((float)$body['data']['amount'], 5))
-        ];
+        return new CurrencyProviderDto($targetCurrency, round((float)$body['data']['amount'], 5));
     }
 }