+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.3.0] - 2024-06-19
 - add `coinbase.com` as exchange rate provider
+- add chain for currency providers and retries
 
 ## [1.2.0] - 2024-06-14
 - transition to hexagonal architecture